アニメーション動作確認、アニメーションスピード適用まで

+++ ComponentFramework/SorceFiles/SubSystem/Texture.cpp
@@ -3,11 +3,11 @@
 #include	"stb_image.h"
 #include	"../Renderer.h"
 
-Texture::Texture(bool loopFlg, int _cutU, int _cutV, int _anmSpeed)
+Texture::Texture(bool loopFlg, float _cutU, float _cutV, int _anmSpeed)
 	: loop_flg_(loopFlg), m_cutU(_cutU), m_cutV(_cutV), m_anmSpeed(_anmSpeed)
 {
-	m_numU = 0;
-	m_numV = 0;
+	m_numU = 0.0f;
+	m_numV = 0.0f;
 }
 
 // テクスチャをロード
